feat: 计算地形网格

+++ houtu-scene/src/height_map_terrain/mod.rs
@@ -5,15 +5,16 @@ pub use create_mesh_job::*;
 // mod create_mesh_job;
 pub struct HeightmapTerrainData {
     pub buffer: Vec<u8>,
-    pub width: i64,
-    pub height: i64,
+    pub width: i32,
+    pub height: i32,
 }
+#[derive(Debug, Clone, Copy, PartialEq)]
 pub struct HeightmapTerrainStructure {
     pub heightScale: f64,
     pub heightOffset: f64,
-    pub elementsPerHeight: i64,
-    pub stride: i64,
-    pub elementMultiplier: f64,
+    pub elementsPerHeight: i32,
+    pub stride: i32,
+    pub elementMultiplier: i32,
     pub isBigEndian: bool,
     pub lowestEncodedHeight: f64,
     pub highestEncodedHeight: f64,
@@ -25,7 +26,7 @@ impl Default for HeightmapTerrainStructure {
             heightOffset: 0.0,
             elementsPerHeight: 1,
             stride: 1,
-            elementMultiplier: 256.0,
+            elementMultiplier: 256,
             isBigEndian: false,
             lowestEncodedHeight: 0.0,
             highestEncodedHeight: 256.0,
